+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Configurable keystrokes before and after ambient paste.
+- Configurable keystrokes and delays before and after ambient paste.
 
 ### Fixed